fix Search_For_Node return type
